change int to bigint

+++ database/migrations/2024_06_28_082954_create_ppdbs_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('status')->default('baru');
             $table->string('nama_lengkap');
-            $table->integer('nik');
+            $table->bigInteger('nik');
             $table->string('jenis_kelamin');
             $table->string('tempat_lahir');
             $table->date('tanggal_lahir');
